restructured the application by adding controllers and models
